requirements.txt
numpy
psycopg2-binary
pytz
requests
python-dateutil
